Convert from UINT16 to unsigned int for balance
Issue #119
-e 
Original/build commit ID: 9436e3e4654e6b5ce1b9d1321127e2687be2289f

+++ balance.c
@@ -19,9 +19,9 @@
  *
  * @returns 1 if the user has enough balance. 0 if not.
  */
-BOOLEAN has_money(UINT16 amount_h, UINT16 amount_l)
+BOOLEAN has_money(unsigned int amount_h, unsigned int amount_l)
 {
-    UINT16 remainder = amount_h;
+    unsigned int remainder = amount_h;
 
     // If lower balance doesn't contain enough,
     // add 1 to required amount in higher int
@@ -50,9 +50,9 @@ BOOLEAN has_money(UINT16 amount_h, UINT16 amount_l)
  *
  * @returns 1 if the amount was removed. 0 if not.
  */
-BOOLEAN remove_money(UINT16 amount_h, UINT16 amount_l)
+BOOLEAN remove_money(unsigned int amount_h, unsigned int amount_l)
 {
-    UINT16 remainder = amount_h;
+    unsigned int remainder = amount_h;
 
     // If lower balance doesn't contain enough,
     // add 1 to required amount in higher int
@@ -97,9 +97,9 @@ BOOLEAN remove_money(UINT16 amount_h, UINT16 amount_l)
  * @param amount_h Upper 16-bit value of the amount to add
  * @param amount_l Lower 16-bit value of the amount to add
  */
-void add_money(UINT16 amount_h, UINT16 amount_l)
+void add_money(unsigned int amount_h, unsigned int amount_l)
 {
-    UINT16 overflow = amount_h;
+    unsigned int overflow = amount_h;
 
     // Check for overflow in lower
     if (game_state.balance[0] != 0 && amount_l > (0xFFFFU - game_state.balance[0]))
